12/10 Prototype.py
Fixed the issue with functions not being able to used on self by using another parameter. Now it looks for user and target. as well as now the when the user makes a move they have to pick the number from their deck.

Also now starting on getting all the numbers to be moves in CLI like for the shop.

ADDED shield, shield resets after opponents turn.

NEED to add print that displays cards, deck number for cards, and i want to add levels to cards so ill make them also display card level. Card levels will come from buying duplicates in store.