12/10 FindableFunctions.py
ADDED shield mechanic that will block the ATTACK's damage, but only absorb damage from the HEAVY. BLOCK adds shield thats removed at the end of the opponents turn. Changed all functions parameters so that the func() will take in the user and target and which ever one is needed in the findable functions will use which ever they need

PLANS, add a level parameter which is multiplied by 5 and added to the damage, shield, or heal.  Also plans on adding mana so shield and heal are not the same and ones just temp. 
ALSO PLANS to add a SLASH function which will damage multiple enemies but will distribute damage. And plans to add elemental damage types and enemies
ALSO need to make a check shield function so that attacking functions wont have redundant code.