12/11 Prototype.py
FIXED SO MUCH 
NOW the deck prints the move names and not numbers which made it confusing to know if the number was the move or the option. 
SHOP NOW SHOWS MOVE NAMES
SHOP NOW ONLY GIVES CARDS FROM THE SHOP AND YOU CANT JUST GET WHAT EVER MOVE

NEED to work on getting the level system maybe by doing a nested dict.