12/12 FindableFunctions.py
Finished 5 moves, added a new function for checking shield for all future offense functions. Also added emojis in the prints for distinction and LEVELING SYSTEM that also works with the printed numbers. ADDED "chance" which will have a chance to miss. And added card level function to count cards for level system.